Initial Versioning of Elasticsearch/OpenSearch config (#7)
As we expand this deployment to support multiple versions of ES/OS we will require customizing the config.yaml as that cluster requires. This lays some initial groundwork with an ES_6 and ES_7 config classes. I imagine we may want to add minor versions in the future with the mindset that these major version config classes will remain and be the default for the major version with a minor version taking priority (e.g. ES_6_8) if that cluster version is used.

Signed-off-by: Tanner Lewis <lewijacn@amazon.com>